Added navigation bar support for register slot and view bill activity
